+++ src/Additions/Repo_List_Table.php
@@ -138,8 +138,7 @@ class Repo_List_Table extends \WP_List_Table {
 	 * @return string Text to be placed inside the column <td> (site title only)
 	 **************************************************************************/
 	public function column_slug( $item ) {
-		// phpcs:disable WordPress.Security.NonceVerification.Recommended
-		// phpcs:disable WordPress.Security.ValidatedSanitizedInput
+		// phpcs:disable WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput
 		$page = isset( $_REQUEST['page'] ) ? sanitize_title_with_dashes( wp_slash( $_REQUEST['page'] ) ) : null;
 		$tab  = isset( $_REQUEST['tab'] ) ? sanitize_title_with_dashes( wp_slash( $_REQUEST['tab'] ) ) : null;
 		// phpcs:enable
@@ -154,7 +153,7 @@ class Repo_List_Table extends \WP_List_Table {
 		// Build row actions.
 		$actions = [
 			// 'edit'   => sprintf( '<a href="%s&action=%s&slug=%s">Edit</a>', $location, 'edit', $item['ID'] ),
-			'delete' => sprintf( '<a href="%s&action=%s&slug=%s">Delete</a>', wp_nonce_url( $location, -1, '_wpnonce_row_action_delete' ), 'delete', $item['ID'] ),
+			'delete' => sprintf( '<a href="%s&action=%s&slug=%s">Delete</a>', wp_nonce_url( $location, 'delete_row_item', '_wpnonce_row_action_delete' ), 'delete_row_item', $item['ID'] ),
 		];
 
 		// Return the title contents.
@@ -270,47 +269,23 @@ class Repo_List_Table extends \WP_List_Table {
 	 **************************************************************************/
 	public function process_bulk_action() {
 		// Detect when a bulk action is being triggered...
-		if ( 'delete' === $this->current_action() ) {
-			$this->check_nonce();
-			// phpcs:ignore WordPress.Security
-			$slugs = isset( $_REQUEST['slug'] ) ? wp_unslash( $_REQUEST['slug'] ) : null;
-			$slugs = is_array( $slugs ) ? $slugs : (array) $slugs;
-			foreach ( $slugs as $slug ) {
-				foreach ( self::$options as $key => $option ) {
-					if ( in_array( $slug, $option, true ) ) {
-						unset( self::$options[ $key ] );
-						update_site_option( 'git_updater_additions', self::$options );
-					}
+		if ( ! isset( $_REQUEST['_wpnonce_row_action_delete'] )
+				|| ! \wp_verify_nonce( \sanitize_key( \wp_unslash( $_REQUEST['_wpnonce_row_action_delete'] ) ), 'delete_row_item' )
+			) {
+			return;
+		}
+		$slugs = isset( $_REQUEST['slug'] ) ? wp_unslash( $_REQUEST['slug'] ) : null;
+		$slugs = is_array( $slugs ) ? $slugs : (array) $slugs;
+		foreach ( $slugs as $slug ) {
+			foreach ( self::$options as $key => $option ) {
+				if ( in_array( $slug, $option, true ) ) {
+					unset( self::$options[ $key ] );
+					update_site_option( 'git_updater_additions', self::$options );
 				}
 			}
 		}
 		if ( 'edit' === $this->current_action() ) {
 			wp_die( esc_html__( 'Items would go to edit when we write that code.', 'git-updater-additions' ) );
-		}
-	}
-
-	/**
-	 * Check nonces.
-	 *
-	 * @return void
-	 */
-	private function check_nonce() {
-		$nonce_exists = false;
-		$nonces       = [
-			'_wpnonce_list'              => 'process-items',
-			'_wpnonce_row_action_delete' => -1,
-		];
-		foreach ( $nonces as $nonce => $action ) {
-			if ( array_key_exists( $nonce, $_REQUEST ) ) {
-				$nonce_exists = true;
-				break;
-			}
-		}
-		if ( ! $nonce_exists ) {
-			wp_die( esc_html__( 'A nonce was not properly set. Please report an issue.', 'git-updater-additions' ) );
-		}
-		if ( ! wp_verify_nonce( sanitize_key( wp_unslash( $_REQUEST[ $nonce ] ) ), $action ) ) {
-			wp_die( esc_html__( 'Your nonce did not verify.', 'git-updater-additions' ) );
 		}
 	}