+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* update nonce check in `class Repo_List_Table`
 
 #### 5.5.1 / 2022-02-05
 * add nonce check for saving settings