+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* use `sanitize_key()` for nonces
 * update nonce check in `class Repo_List_Table`
 
 #### 5.5.1 / 2022-02-05