+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 5.5.2 / 2022-02-08
 * use `sanitize_key()` for nonces
 * update nonce check in `class Repo_List_Table`